# api/swagger.yaml
openapi: 3.0.3
info:
  title: Athlete Hub API
  description: |
    REST API for the Athlete Hub sports events platform.

    ## Authentication
    - **Protected routes** require `Authorization: Bearer <session_token>`
    - **Public routes** require no authentication

    ## Base URL
    All endpoints are prefixed with `/api`.
  version: "1.0.0"
  contact:
    name: Athlete Hub Dev
servers:
  - url: /api
    description: Same-origin (development & production)
...
paths:
...
  /public/teams:
    get:
      tags: [Marketplace]
      summary: Browse public athlete Triboos (teams)
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: sort
          in: query
          schema:
            type: string
            enum: [members, newest]
      responses:
        "200":
          description: Paginated public teams list
  /public/teams/{slug}:
    get:
      tags: [Marketplace]
      summary: Public Triboo profile by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Team profile with member preview
        "404":
          description: Not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT